+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Use of external EEPROM
+- Clearing punches in the Android app
 
 ## [2.1.0] - 2025.11.30